Move If function to separate file
