+++ lib/keisan/ast/function.rb
@@ -71,34 +71,16 @@ module Keisan
         "#{name}(#{children.map(&:to_s).join(',')})"
       end
     end
+  end
+end
 
-    class If < Function
-      def value(context = nil)
-        unless (2..3).cover? children.size
-          raise Keisan::Exceptions::InvalidFunctionError.new("Require 2 or 3 arguments to if")
-        end
+require_relative "functions/if"
 
-        bool = children[0].value(context)
-
-        if bool
-          children[1].value(context)
-        else
-          children.size == 3 ? children[2].value(context) : nil
-        end
-      end
-
-      def unbound_functions(context = nil)
-        context ||= Keisan::Context.new
-
-        children.inject(Set.new) do |res, child|
-          res | child.unbound_functions(context)
-        end
-      end
-    end
-
+module Keisan
+  module AST
     class Function
       BUILD_CLASSES = {
-        "if" => If
+        "if" => AST::Functions::If
       }.freeze
 
       def self.build(name, arguments = [])
